backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib
bcrypt==4.0.1
websockets
scikit-learn
numpy
pandas
python-multipart
